# Benchmark in-memory, index-only search