--- a/cmd/frontend/db/repos_mock.go
+++ b/cmd/frontend/db/repos_mock.go
@@ -44,7 +44,7 @@ func (s *MockRepos) MockGet_Return(t *testing.T, returns *types.Repo) (called *b
 	return
 }
 
-func (s *MockRepos) MockGetByName(t *testing.T, want api.RepoName, repo api.RepoID) (called *bool) {
+func (s *MockRepos) MockGetByName(t testing.TB, want api.RepoName, repo api.RepoID) (called *bool) {
 	called = new(bool)
 	s.GetByName = func(ctx context.Context, name api.RepoName) (*types.Repo, error) {
 		*called = true
@@ -57,7 +57,7 @@ func (s *MockRepos) MockGetByName(t *testing.T, want api.RepoName, repo api.Repo
 	return
 }
 
-func (s *MockRepos) MockList(t *testing.T, wantRepos ...api.RepoName) (called *bool) {
+func (s *MockRepos) MockList(t testing.TB, wantRepos ...api.RepoName) (called *bool) {
 	called = new(bool)
 	s.List = func(ctx context.Context, opt ReposListOptions) ([]*types.Repo, error) {
 		*called = true