--- a/cmd/frontend/graphqlbackend/search.go
+++ b/cmd/frontend/graphqlbackend/search.go
@@ -200,7 +200,7 @@ func getSampleRepos(ctx context.Context) ([]*types.Repo, error) {
 
 // resolveRepositories calls doResolveRepositories, caching the result for the common
 // case where effectiveRepoFieldValues == nil.
-func (r *searchResolver) resolveRepositories(ctx context.Context, effectiveRepoFieldValues []string, includeResolvers bool) (repoRevs, missingRepoRevs []*search.RepositoryRevisions, repoResults []*searchSuggestionResolver, overLimit bool, err error) {
+func (r *searchResolver) resolveRepositories(ctx context.Context, effectiveRepoFieldValues []string) (repoRevs, missingRepoRevs []*search.RepositoryRevisions, repoResults []*searchSuggestionResolver, overLimit bool, err error) {
 	tr, ctx := trace.New(ctx, "graphql.resolveRepositories", fmt.Sprintf("effectiveRepoFieldValues: %v", effectiveRepoFieldValues))
 	defer func() {
 		if err != nil {
@@ -243,7 +243,6 @@ func (r *searchResolver) resolveRepositories(ctx context.Context, effectiveRepoF
 		onlyArchived:     archived == Only || archived == True,
 		noArchived:       archived == No || archived == False,
 		commitAfter:      commitAfter,
-		includeResolvers: includeResolvers,
 	})
 	tr.LazyPrintf("resolveRepositories - done")
 	if effectiveRepoFieldValues == nil {
@@ -358,7 +357,6 @@ type resolveRepoOp struct {
 	noArchived       bool
 	onlyArchived     bool
 	commitAfter      string
-	includeResolvers bool
 }
 
 func resolveRepositories(ctx context.Context, op resolveRepoOp) (repoRevisions, missingRepoRevisions []*search.RepositoryRevisions, repoResolvers []*searchSuggestionResolver, overLimit bool, err error) {
@@ -425,9 +423,6 @@ func resolveRepositories(ctx context.Context, op resolveRepoOp) (repoRevisions, 
 	overLimit = len(repos) >= maxRepoListSize
 
 	repoRevisions = make([]*search.RepositoryRevisions, 0, len(repos))
-	if op.includeResolvers {
-		repoResolvers = make([]*searchSuggestionResolver, 0, len(repos))
-	}
 	tr.LazyPrintf("Associate/validate revs - start")
 	for _, repo := range repos {
 		repoRev := &search.RepositoryRevisions{Repo: repo}
@@ -478,15 +473,8 @@ func resolveRepositories(ctx context.Context, op resolveRepoOp) (repoRevisions, 
 		}
 
 		repoRevisions = append(repoRevisions, repoRev)
-
-		if op.includeResolvers {
-			repoResolver := &repositoryResolver{repo: repo}
-			repoResolvers = append(repoResolvers, newSearchResultResolver(
-				repoResolver,
-				math.MaxInt32,
-			))
-		}
 	}
+
 	tr.LazyPrintf("Associate/validate revs - done")
 
 	if op.commitAfter != "" {
@@ -556,7 +544,7 @@ func optimizeRepoPatternWithHeuristics(repoPattern string) string {
 }
 
 func (r *searchResolver) suggestFilePaths(ctx context.Context, limit int) ([]*searchSuggestionResolver, error) {
-	repos, _, _, overLimit, err := r.resolveRepositories(ctx, nil, false)
+	repos, _, _, overLimit, err := r.resolveRepositories(ctx, nil)
 	if err != nil {
 		return nil, err
 	}