… cmd/frontend/graphqlbackend/search_repositories_test.go
… cmd/frontend/graphqlbackend/search_results_test.go
… cmd/frontend/graphqlbackend/textsearch_test.go
… cmd/frontend/internal/pkg/search/search.go