@mrnugget 2019-06-27
Benchmark as of cb8cc7f:

```
$ cd cmd/frontend/graphqlbackend && go test -run ^$ -bench=SearchResults -count 15 -benchmem > ~/old.txt
$ cat ~/old.txt
goos: darwin
goarch: amd64
pkg: github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend
BenchmarkSearchResults-8             300           4509566 ns/op         2917818 B/op      30717 allocs/op
BenchmarkSearchResults-8             300           4475652 ns/op         2911605 B/op      30714 allocs/op
BenchmarkSearchResults-8             300           4429977 ns/op         2907353 B/op      30713 allocs/op
BenchmarkSearchResults-8             300           4433480 ns/op         2904842 B/op      30712 allocs/op
BenchmarkSearchResults-8             300           4443603 ns/op         2903430 B/op      30711 allocs/op
BenchmarkSearchResults-8             300           4447677 ns/op         2901335 B/op      30710 allocs/op
BenchmarkSearchResults-8             300           4492250 ns/op         2900874 B/op      30710 allocs/op
BenchmarkSearchResults-8             300           4452620 ns/op         2900068 B/op      30710 allocs/op
BenchmarkSearchResults-8             300           4451069 ns/op         2899845 B/op      30710 allocs/op
BenchmarkSearchResults-8             300           4431784 ns/op         2899214 B/op      30709 allocs/op
BenchmarkSearchResults-8             300           4418570 ns/op         2899303 B/op      30709 allocs/op
BenchmarkSearchResults-8             300           4433587 ns/op         2898367 B/op      30709 allocs/op
BenchmarkSearchResults-8             300           4481626 ns/op         2898448 B/op      30709 allocs/op
BenchmarkSearchResults-8             300           4429456 ns/op         2898096 B/op      30709 allocs/op
BenchmarkSearchResults-8             300           4453930 ns/op         2898433 B/op      30709 allocs/op
PASS
ok      github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend  27.351s
```