@mrnugget 2019-06-27
State as of 70fd9e4:

```
λ cat ~/new.txt
goos: darwin
goarch: amd64
pkg: github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend
BenchmarkSearchResults-8             300           4090711 ns/op         2313145 B/op      20714 allocs/op
BenchmarkSearchResults-8             300           4000503 ns/op         2306827 B/op      20712 allocs/op
BenchmarkSearchResults-8             500           3967640 ns/op         2302278 B/op      20710 allocs/op
BenchmarkSearchResults-8             300           3997998 ns/op         2300599 B/op      20710 allocs/op
BenchmarkSearchResults-8             300           3989411 ns/op         2300495 B/op      20709 allocs/op
BenchmarkSearchResults-8             300           4090362 ns/op         2298948 B/op      20709 allocs/op
BenchmarkSearchResults-8             300           3971854 ns/op         2298050 B/op      20708 allocs/op
BenchmarkSearchResults-8             300           4099103 ns/op         2297548 B/op      20708 allocs/op
BenchmarkSearchResults-8             300           3976969 ns/op         2297904 B/op      20708 allocs/op
BenchmarkSearchResults-8             500           3958725 ns/op         2297140 B/op      20708 allocs/op
BenchmarkSearchResults-8             300           3956273 ns/op         2297258 B/op      20708 allocs/op
BenchmarkSearchResults-8             500           3984142 ns/op         2296310 B/op      20708 allocs/op
BenchmarkSearchResults-8             500           3988056 ns/op         2296664 B/op      20708 allocs/op
BenchmarkSearchResults-8             300           4076461 ns/op         2296260 B/op      20708 allocs/op
BenchmarkSearchResults-8             500           3962351 ns/op         2295541 B/op      20708 allocs/op
PASS
ok      github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend  28.640s
```
Comparing cb8cc7f (`old.txt`) to 70fd9e4 (`new.txt):

```
$ benchstat ~/old.txt ~/new.txt
name             old time/op    new time/op    delta
SearchResults-8    4.45ms ± 1%    4.01ms ± 2%   -9.99%  (p=0.000 n=15+15)

name             old alloc/op   new alloc/op   delta
SearchResults-8    2.90MB ± 0%    2.30MB ± 0%  -20.80%  (p=0.000 n=14+13)

name             old allocs/op  new allocs/op  delta
SearchResults-8     30.7k ± 0%     20.7k ± 0%  -32.57%  (p=0.000 n=14+14)